[BUGFIX] Delete move placeholders upon record deletion
If a record was moved in a draft workspace and subsequently
removed in the live workspace, then the move placeholder would
remain and show up as an actual record in the backend. To the
end user it appears as if the record was not deleted as one
would now see the move-placeholder.

If the mentioned moved record would be removed instead in the
very same draft workspace, the move-to-pointer was removed
correctly, but the move-placeholder stayed as well.

Resolves: #39383
Releases: master, 6.2
Change-Id: I3b0593332650795539c2d7eac075515dbd981b45
Reviewed-on: http://review.typo3.org/13496
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>