[BUGFIX] Distinguish unassigend columns and colPos 0
When using backend layout columns without a colPos value they should be
just placeholders with the label "not assigned". Currently they are
showing the content of the column 0 instead if there is such a column in
the backend layout.

The label "not assigned" is used for columns without any
label, otherwise the label is used together with the suffix
"(not assigned)".

Change-Id: I02c418eebdd9345c3066aa8c3eeec353d2cd9e58
Resolves: #25157
Resolves: #45550
Releases: 4.5, 4.7, 6.0, 6.1, 6.2
Reviewed-on: https://review.typo3.org/18349
Reviewed-by: Carsten Lausen
Tested-by: Carsten Lausen
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Mattias Nilsson
Tested-by: Mattias Nilsson
Reviewed-by: Stefan Neufeind