[TASK] Revamp EXT:recycler
Refactors the recycler extension to a modern architecture. The backend
is based on Extbase and Fluid, the UI is based on jQuery and
Twitter Bootstrap now.

Due to restrictions in the core, non-admin users cannot restore deleted
pages for now.

Kudos to Felix Kopp for porting the base to Extbase and Fluid.

Releases: master
Resolves: #64420
Change-Id: I9d330981af0b42703b8352c1d61bec818e08b38e
Reviewed-on: http://review.typo3.org/36109
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>