[TASK] Clean up TCEforms main table and according CSS definitions
TCEforms has a major table around each editing form element.

In the current version, it creates an ugly background, plus some
wrong borders. This is now replaced by a div container and
some cleanups. Additionally, the labels inside a palette are
ordered now in a better way as well.

Releases: 6.2
Resolves: #49596
Change-Id: Ifa46ce40ff4437b2bb3e7d32e5e4e56caba0caec
Reviewed-on: https://review.typo3.org/21802
Reviewed-by: Wouter Wolters
Tested-by: Wouter Wolters
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn