[BUGFIX] Enable add markers in LoginController signal
Currently its not possible to have signales handle parameter of slots
as a reference. Because of this its not possible to add marker in the
emit. With this patch the parameter containing the markers is taken
after process and then the markers get handed over to the reference
again.

Resolves: #51892
Releases: 6.2
Change-Id: Id6f92195608ca8684fdbfe911b193789055d6775
Reviewed-on: https://review.typo3.org/23728
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring