[FEATURE] Make getPlaceholderValue recursive for __row
If __row|uid_foreign|field is a reference its not evaluated
recursive the placeholder shows a uid. With this patch the
content of the field is shown.

Additonally the type of the metadata field of sys_file was changed
to inline because select does not support the foreign_field property.

Resolves: #52630
Releases: 6.2
Change-Id: Iee30274d2e94f73e877b761d119801bb205689c1
Reviewed-on: https://review.typo3.org/24537
Reviewed-by: Oliver Hader
Tested-by: Oliver Hader