[BUGFIX] Use getPagesTSconfig($tscPID) only once per ID
PageLayoutView::isRTEforField is called on every record that is
displayed in a Web>Page. And this will call
BackendUtility::getPagesTSconfig($tscPID)
every time once and again,
and thus parsing the PageTS over and over on the same hit.
If you have 50 content elements on a page,
this can get very expensive.

Solution: Create a variable that stores the RTEsetup for
 a particular ID

Resolves: #53368
Releases: 6.2

Change-Id: I7efcefb0c12c856e9417ca8bb655acaabd4f15cc
Reviewed-on: https://review.typo3.org/25181
Reviewed-by: Oliver Klee
Reviewed-by: Jo Hasenau
Reviewed-by: Alexander Opitz
Tested-by: Alexander Opitz
Reviewed-by: Sebastian Fischer
Reviewed-by: Wouter Wolters
Reviewed-by: Peter Foerger
Tested-by: Peter Foerger
Tested-by: Philipp Gampe
Reviewed-by: Philipp Gampe
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind