[BUGFIX] Use default BE layout if current not available
If a specific backend layout is selected in the page properties
and the backend layout is removed later on, the page module is
empty and shows nothing. As a solution the default layout should
be loaded like it is the case if no BE layout is defined at all.

Resolves: #53608
Releases: 6.2
Change-Id: I192833aacc4785a7dfd8e1fb5ed34f12e33d4f5c
Reviewed-on: https://review.typo3.org/25374
Reviewed-by: Philipp Gampe
Tested-by: Philipp Gampe
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind