[BUGFIX] Page module: Allow to paste in empty columns
Since quite a long time it was not possible to paste a content
element into an empty column of the page module. This patch
introduces an icon in the top right corner of each column if
a tt_content element is on the clipboard (Default pad). When
the icon gets clicked the content elements from the clipboard
get moved/copied into the selected column/language.

Resolves: #15080
Releases: 6.2, 6.1
Change-Id: If52905446eb11c268d0fee83f150efae0945fa29
Reviewed-on: https://review.typo3.org/25532
Reviewed-by: Markus Klein
Tested-by: Markus Klein
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind