[CLEANUP] Remove old tt_content DB fields
With TYPO3 4.0 and the new CSS styled content
and the new RTE HTMLArea, some fields were still
included in the core, but disabled by default as
soon as CSS Styled Content was installed. If not
specifically activated, the fields were disabled
in the last 10 versions and are never rendered
anymore by default via CSS Styled Content.

The DB fields are:
  * text_align
  * text_face
  * text_size
  * text_color
  * text_properties

As they are not in use at all (since a long time)
the patch removes any info about these fields
without further notice or any upgrade wizard.

If somebody needs the fields, he/she needs
to create an own extension with the fields.

Releases: 6.2
Resolves: #55479
Change-Id: I77655cec0a27903a0a1df017808f9603564ef739
Reviewed-on: https://review.typo3.org/27168
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Christian Kuhn
Reviewed-by: Marcin Sągol
Reviewed-by: Georg Ringer
Tested-by: Georg Ringer