[BUGFIX] IRRE bidirectional relation not highlighted
Currently the user is not able to recognize that he is working on
record relations. These records may be used in several places in
the system and the user should be aware of that specific behaviour.

In previous versions of TYPO3 there has been a wrapper which
highlighted input fields, but had no explanation. This was removed
later. To restore the functionality and make it understandable to the
user why he should edit the fields wisely, the wrapper has been replaced
by a flash message. This explains to the user what happens when the
record gets changed.

Resolves: #55553
Releases: 6.2
Change-Id: Ic8a41cdb251cfc8d0134e2b53927ff71ddd5e594
Reviewed-on: https://review.typo3.org/27235
Reviewed-by: Stefan Froemken
Tested-by: Stefan Froemken
Reviewed-by: Markus Klein
Tested-by: Markus Klein