[TASK] FormEngine: "field has changed" icon via CSS
When editing a record in FormEngine, each field that is
modified gets an arrow symbol in front of the
input/select field.

There are certain downsides to this implementation:
- Having images and hard-coded srcs in each field
clutters the HTML and has ugly JS to change the field
- it's hard to style / exchange the icon

The new solution does the replacement via CSS classes.
The CSS class just adds a blue background to the
changed form field or does a color change
to the checkbox label currently.

Releases: master
Resolves: #38752
Change-Id: I94b6cc8550f7539961aa7ee97a8081e7d0ae5931
Reviewed-on: http://review.typo3.org/12708
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>