[BUGFIX] Different button behaviour Page module views
In the Page module the button behaviour for adding new content elements
"Add a new record at this place" is different in the "columns" and
"languages" view.

The columns view shows the all buttons in a column when hovering over this
column, the languages view will show one button when hovering over this
hidden button. A backend user needs to know where the button is.

The languages view should have the same behaviour as the columns view.

Change-Id: Ia1b98de5b8622c7586bd03e7769ba4aa39121e38
Resolves: #56722
Releases: 6.2, 6.1
Related: #56720
Reviewed-on: https://review.typo3.org/28219
Reviewed-by: Steffen Ritter
Tested-by: Steffen Ritter
Reviewed-by: Stefan Neufeind
Tested-by: Stefan Neufeind