[TASK] Remove not maintainable test
ext:backend test getNodesSetsIsMountPointField() works on getNode()
of PageTree DataProvider. The method has tons of dependencies that
are sometimes even static, and the method is recursive. It is very
hard to test properly and the current test is still not mocking
completly. The test is removed for now since it does not give us
much benefit and is hard to maintain in current state, the subject
method should be covered by tests again if it was refactored.

Change-Id: I344c584b4d6e3868471f4857a08223be3a5d03e4
Resolves: #56907
Related: #51436
Releases: 6.2
Reviewed-on: https://review.typo3.org/28384
Reviewed-by: Wouter Wolters
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Christian Kuhn
Tested-by: Christian Kuhn