[BUGFIX] Enable backend login for Internet Explorer 8 users
Due to some mystical IE8 behavior the global TYPO3 object gets reset
if any "var" occurs in Javascript. It is not necessary if its called or
not but it seems IE8 creates a new object. This patch adds the global
initialization of the TYPO3 object by using window[] syntax.

Resolves: #56988
Releases: 6.2
Change-Id: I5ce246790d1e3a1e388f106d085c96dae4e4304e
Reviewed-on: https://review.typo3.org/28929
Reviewed-by: Alexander Opitz
Tested-by: Alexander Opitz
Reviewed-by: Markus Klein
Tested-by: Markus Klein