[TASK] Trigger events in DragUploader
The DragUploader implementation prevents most of its JavaScript
API to be interacted with from outside the module. Especially
there is no public event allowing interaction based on ajax
responses such as "uploadComplete" or "uploadError".

This patch adds a couple of events being triggered on the $trigger
property of the DragUploader module. That's meant to be the "add"
button in DragUploader context, so it's a mandatory DOM node.

The introduced events are:
* uploadStart
* uploadError
* updateProgress
* uploadComplete

Change-Id: I983d5528247f88d319c2385e3fb032cbd701fa33
Resolves: #57454
Releases: 6.2
Reviewed-on: https://review.typo3.org/28976
Reviewed-by: Stefan Neufeind
Reviewed-by: Helmut Hummel
Tested-by: Stefan Neufeind
Reviewed-by: Wouter Wolters
Tested-by: Frans Saris
Reviewed-by: Frans Saris
Reviewed-by: Xavier Perseguers
Tested-by: Xavier Perseguers