[FEATURE] Configurable width of the Element Browser Popups
The width of the element browser is sadly hardcoded to 600 and 650.

In many situations the element browser would need more place and
with this tiny change it would be easily possible to change
that behaviour without overloading the whole file.

The patch allows to set dimensions on a per-user basis via
userTSconfig.

Resolves: #58031
Resolves: #20929
Resolves: #22976
Releases: master
Change-Id: Ida3d01c7fd0457b4a88a33a10b0c19792ba9751a
Reviewed-on: http://review.typo3.org/29618
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Alexander Opitz <opitz.alexander@googlemail.com>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>