[TASK] Enhance localization behavior of pages.media
With issue #57488 the pages.media field became localizable again,
however the selective approach turns out to be problematic in
terms of overlaying data in the frontend (the "mergeIfNotBlank"
settings and the like).
Thus, instead of letting the editor localize the references after
persisting the parent and having to overlay/fallback in the
frontend, now the references are already shown in the edit view
during the localization (still before being persisted).

Resolves: #58476
Releases: 6.2
Change-Id: Id6980f2a5cfea109375afce27f62cad8c101c58e
Reviewed-on: https://review.typo3.org/29839
Reviewed-by: Anja Leichsenring
Tested-by: Anja Leichsenring
Reviewed-by: Marc Bastian Heinrichs
Tested-by: Marc Bastian Heinrichs