[BUGFIX] Request update broken for select fields
For select-type fields displayed as two side-by-side selectors
which are registered as "requestUpdate" fields, the form submission
happens before the value is copied to the left-hand selector.
The form is thus saved without value.

Full description in the bug report, which also contains a
test extension and steps to reproduce.

Resolves: #58755
Releases: 6.2
Change-Id: Ie94def723537202287655c650cee7dfa09815f4f
Reviewed-on: https://review.typo3.org/30044
Reviewed-by: Peter Foerger
Reviewed-by: Fabien Udriot
Tested-by: Fabien Udriot
Reviewed-by: Tomas Norre Mikkelsen
Reviewed-by: Lorenz Ulrich
Tested-by: Lorenz Ulrich
Reviewed-by: Markus Klein
Tested-by: Markus Klein