[BUGFIX] Use delayedImportElement after uploading files
When uploading multiple files in TCEForm the DrapUploader
adds all uploaded files directly to the inline element of the
form. But when a previous request isn't finished yet, it
breaks the processing of last request.

This patch changes the use of importElement to
delayedImportElement so the next element is only added
when the previous is processed.

Resolves: #58956
Resolves: #57472
Resolves: #57190
Releases: 6.2
Change-Id: Iccad2ee19a5462cb63ed749ad60cb685460eb66d
Reviewed-on: https://review.typo3.org/30609
Reviewed-by: Helmut Hummel
Tested-by: Marc Fell
Reviewed-by: Markus Klein
Tested-by: Markus Klein