[CLEANUP] Move test file deletion into UnitTestCase class
In order to unify removal of files and directories created by unit
tests, the basic UnitTestCase class get a property to register
created files and directories. Removing those files is
implemented in tearDown().

Tests now just need to add created files to $this->testFilesToDelete
to get the automatically cleaned up after test run. The method is
constructed to only remove stuff within typo3conf/ext and typo3temp
to reduce the risk of removing too, it throws an exception if this
rule is not followed.

Change-Id: Ibca2b3a006432f2335fff8f0b0c05c62e363878c
Resolves: #60123
Releases: 6.3
Reviewed-on: http://review.typo3.org/31359
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>