[TASK] Clean up backend login controller Part 1
To prepare a switch of the backend login view to a fluid rendered
template, the current controller is first simlpfied a bit and some
ancient code is removed.

With t3skin, the default template of ext:backend/.../login.html is
always overloaded by ext:t3skin/.../login.html. The patch overwrites
the content of the ext:backend template with the content from
ext:login and removes the template and overload registration in
ext:login.

The old template in ext:backend was not used for a very long time
but includes some old handling like loginBoxImage_rotationFolder
encapsulated in makeLoginBoxImage(). This code is dropped without
substitution.

References to two ancient login images are dropped with the removal
of makeLoginBoxImage() and can be finally removed along the way.

Change-Id: I057fbe062b138ab4f428e49e511bc46363a8a7ed
Resolves: #60559
Releases: 6.3
Reviewed-on: http://review.typo3.org/31802
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>