[FEATURE] BE: Use proper placeholder shim for IE <= 9
The current solution to add HTML5 placeholder
attribute support for unsupported browsers
(which are IE9 and lower) is using a combination
of ExtJS and Prototype, which both need
to be loaded for that.

Currently this is used in FormEngine and
the backend login form.

However, there are better ways of creating
these fallbacks, one being Placeholder.JS
(see http://jamesallardice.github.io/Placeholders.js/)
which can seamlessly be integrated,
and only be loaded for IE9 and less.

Resolves: #60578
Releases: 6.3
Change-Id: If8443276b85d15222c4fc0954e3f79b50ce6a4c9
Reviewed-on: http://review.typo3.org/31830
Reviewed-by: Felix Kopp <felix-source@phorax.com>
Tested-by: Felix Kopp <felix-source@phorax.com>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>