[TASK] Add first documentation of breaking changes
For better documentation of important changes a new file structure
is introduced that substitutes the old NEWS.md file. Some recent
changes are already added with this change as examples, but all
patches since 6.2 release need a post-merge review to see which
further patches need documentation. In future, changelog files for
important changes should be provided together with the patch that
introduces the change.

Resolves: #60186
Resolves: #59741
Resolves: #59740
Releases: master
Change-Id: I6e154262ed49542a5b15e7c68313c9f11dea5fe5
Reviewed-on: http://review.typo3.org/31853
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Ernesto Baschny <ernst@cron-it.de>
Tested-by: Ernesto Baschny <ernst@cron-it.de>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>