[TASK] Rewrite Backend Login JS to use jQuery
The backend login process uses various
JavaScript parts.

1) Some basic redirects on pure JavaScript
that is inline in the LoginController.

2) A flat and hardcoded JS file in the
login.html template which actually
uses ExtJS and Prototype for various
cookie checks and usability checks.

3) RSA Authentication with Backend Login
for doing an AJAX call. This is done in
ExtJS.

All of the parts above are rewritten
using the proper API calls based on
jQuery.

Resolves: #60576
Releases: 6.3
Change-Id: Ie64c4736c79327b816c39da5bba533bcc65cee45
Reviewed-on: http://review.typo3.org/31826
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>