[BUGFIX] Only reload pagetree if pages are changed
When using the click menu to hide, unhide or delete
a record, previously the page tree was refreshed,
although those changes are only reflected in the
pagetree if page properties are changed.

Adapt the click menu to only refresh the tree
if a page record is changed or deleted.

Resolves: #40687
Releases: 6.2, master
Change-Id: I32844b359abd790f1dbd6839eec6feeb6e62bd02
Reviewed-on: http://review.typo3.org/33573
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>