[FEATURE] Load twbs JS and jquery on every BE request
In order to streamline the JavaScript
code and the usage of Twitter Bootstrap
in TYPO3, jQuery and JS of Bootstrap
are included on every request now.

Twitter Bootstrap does not support AMD
modules yet (only with 4.0), so a separate
encapsulated module needs to be built,
which will be updated when bootstrap version
is updated as well.

Releases: master
Resolves: #62877
Change-Id: I4c3185b0c2cc8a57890efdd0687b784bae8c478c
Reviewed-on: http://review.typo3.org/34004
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>