[TASK] Move FormEngine SelectboxFilter to jQuery
One of the last bits and pieces with PrototypeJS
is a relatively new feature introduced with 6.2
for filtering records in a multi-select (two boxes mode).

The JS functionality was completely written,
styling is optimized to fit bootstrap needs.

Through using deferred event handlers, the
PHP code is reduced.

Releases: master
Resolves: #62991
Change-Id: Idb189a871b8499fd5129773641f6ac5e8fb373d2
Reviewed-on: http://review.typo3.org/34212
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>