[!!!][TASK] Migrate non-ExtJS trees to jQuery
Existing tree components that are
not using ExtJS are mainly within
the ElementBrowser and the FolderTree,
which is the navigation component inside filelist.

However, the "old" version of the original pagetree
still exists, and has support for IE5.5 as well, and
an option to filter for items within the pagetree (please
note: this feature was never added to the foldertree).

The old pagetree navigation component itself is now
deprecated (it can still be used in own modules). The
filter functionality is completely removed in the
old pagetree.

The necessary JavaScript code for drag+drop and
expand/collapse of the tree was migrated to RequireJS,
which avoids some more loading of scriptaculous
and prototype.

Releases: master
Resolves: #64059
Change-Id: If8de627d0284d77accdfe55cffe12c591f32df7a
Reviewed-on: http://review.typo3.org/34239
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>