[BUGFIX] jQuery is loaded twice in backend.php
Due to several patches, jQuery is currently
loaded twice in the backend.php script.
Once via window.jQuery (default noConflict)
and via TYPO3.jQuery. The core only uses
window.jQuery in the module menu JS
which is now changed to use TYPO3.jQuery.

As for the current state, as long as jQuery
is not globally available with window.$ due
to the conflict with Prototype, the TYPO3.jQuery
is used throughout the core, and also
in the module menu.

Releases: master
Resolves: #63133
Change-Id: Iadc019ec4b84f2aee833009c1c47449e9c1d2fe7
Reviewed-on: http://review.typo3.org/34455
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>