[TASK] Adapt FormEngine translated records to Bootstrap
When editing a translated record, a message
is shown of what is written in the original field.
Additionally, if the original record was modified
since the last change, a diff is shown. If an icon
is available then it is displayed.

The patch adapts the following things:

1. CSS classes are renamed in order to reflect
FormEngine and not old TCEForms namings.

2. If the original language does not have an icon
the label of the icon is shown again. This actually
a bug introduced when adapting sprites in FormEngine
and is now back to the original version.

3. Styling is adapted, also for the diff to use bootstrap
values.

Resolves: #63440
Releases: master
Change-Id: I314446e3e184dff26f52c70e37c6969c069728bd
Reviewed-on: http://review.typo3.org/34800
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>
Tested-by: Stefan Neufeind <typo3.neufeind@speedpartner.de>