[!!!][TASK] Remove option "compat_version"
The global option
$TYPO3_CONF_VARS[SYS][compat_version]
was introduced for having old frontend output
while upgrading an installation to a newer version.

For example, it was used for conditions within
CSS Styled Content before separate TS templates
for different versions were introduced.

The option mainly works within
GeneralUtility::compat_version, which
is used in several places.

In a normal installation the value equals the
constant TYPO3_branch.

The according update wizard for updating
the option is not needed anymore and is
removed without substitution.

Resolves: #24900
Releases: master
Change-Id: Id3fef21122ad92a9a5a9e90f60179b50d42b6a11
Reviewed-on: http://review.typo3.org/35684
Reviewed-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Tested-by: Tymoteusz Motylewski <t.motylewski@gmail.com>
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Tested-by: Markus Klein <klein.t3@reelworx.at>