[TASK] Remove language "default" from Alternative Page Language and hide record type when not necessary
The pages_language_overlay table has a sys_language_uid
field, but (like all other records) it is filled with "default".
However, due to the concept of two tables for translations
with pages and pages_language_overlay, the default
value does not make any sense and rather causes trouble
within the core when set to "0" / "default".

The patch removes the static entry from the language
dropdown of a pages_language_overlay record.

See also further discussion on forge.

Releases: master
Resolves: #34823
Change-Id: I258a8e2e31411779cc46005d56bc54e2a7a78b7d
Reviewed-on: http://review.typo3.org/35695
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>