[!!!][TASK] FormEngine: Unify resizable textareas
TYPO3 has some special functionality
for FormEngine textareas (without RTE).
The options are available through the users'
settings.

1. Resize a textarea
2. "Autogrow" a textarea. As soon as
the contents are larger than the textarea
the textarea's height is adapted.

The first feature (and option) is dropped. Although
it is working natively in most browsers (except IE)
via CSS, the horizontal resizing is a no-go for forms,
as agreed with the Design Team.

The second feature is rewritten with a unified
solution based on jQuery, which loads only when
needed (and works with AJAX) and works in all
browsers.

Resolves: #64131
Resolves: #55540
Releases: master
Change-Id: Id3f2c520fd76043e2cce80e92732a12e58e395ac
Reviewed-on: http://review.typo3.org/35794
Reviewed-by: Benjamin Kott <info@bk2k.info>
Tested-by: Benjamin Kott <info@bk2k.info>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>