[TASK] Use jQuery UI for Drag&Drop in Page Module
In the process to move away from ExtJS
and fixing some open problems, the JS
code for the drag+drop functionality in the
page module is rewritten based on jQuery UI.

In addition to having a RequireJS module,
the new JS-based DataHandler AJAX call
is used. Also, a content element is only
moved in the HTML code, if the AJAX result
is true. Scrolling works as well, which resolves
the existing problem from forge ticket #50576.

Resolves: #50576
Resolves: #64221
Releases: master
Change-Id: I19cdbf4820d16606125ec908cc7fe4d34df415da
Reviewed-on: http://review.typo3.org/35807
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Benjamin Mack <benni@typo3.org>
Tested-by: Benjamin Mack <benni@typo3.org>