[TASK] Make new content element button in page view speaking
The new content element button is hard to spot for new users
and not self explaining / missleading in the first place.

Unifying the "new record here" button with the corresponding
title from the new record view makes it more clear what kinds
of records will be available to place in a column and at the place
the user has chosen.

Resolves: #65206
Releases: master
Change-Id: Icbe897f943d27771cdbb5e70505d1f47f6adfbdf
Reviewed-on: http://review.typo3.org/37057
Reviewed-by: Michael Oehlhof <typo3@oehlhof.de>
Tested-by: Michael Oehlhof <typo3@oehlhof.de>
Reviewed-by: Nicole Cordes <typo3@cordes.co>
Tested-by: Nicole Cordes <typo3@cordes.co>
Reviewed-by: Jigal van Hemert <jigal.van.hemert@typo3.org>
Tested-by: Jigal van Hemert <jigal.van.hemert@typo3.org>