[TASK] Introduce unified RequestHandling for all requests
The TYPO3 Bootstrap gets a new function run() to unify
existing entrypoints in FE, BE, CLI and Install Tool.

This way the "visible" part of the bootstrap is reduced
to solely using the new "run()" method, which resolves
to a proper RequestHandler depending on the following
constraints.

* Install Tool
* Frontend: eID
* Frontend: TSFE
* Backend: Regular request
* Backend: AJAX
* Command-Line (CLIkeys)

The interface behaves exactly like TYPO3 Flow.

Unlike Flow, the request handlers are currently hardcoded
inside the bootstrap, but can be extended later.

Each RequestHandler logic is currently put from the global
scope into the main "handleRequest()" method.

Further changes are the deprecation of init.php and
the streamlining of Backend requests, as well as diagram
of the bootstrap process.

Resolves: #65437
Releases: master
Change-Id: I3a2ffe60e59005955c5ef3d3c541b61de2b97526
Reviewed-on: http://review.typo3.org/37409
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Helmut Hummel <helmut.hummel@typo3.org>
Tested-by: Helmut Hummel <helmut.hummel@typo3.org>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>