[!!!][FEATURE] Introduce Backend Routing
A new Routing API is introduced in order to streamline
the entrypoints to the TYPO3 Backend.

Instead of using the term "module" for anything linkable
in the backend, the term "routes" fits more. A "module"
or an ajax call is a derivative of a route, which will
build on this foundation.

Routes can be registered via
Configuration/Backend/Routes.php in any extension
and are loaded solely on Backend requests.

There are three new classes:
- Route (a single route with a path and some options)
- Router (API to match paths)
- UriBuilder (Generates a Backend URI)

This patch changes the entrypoint for
login/logout to typo3/index.php?route=...&token=....

The main RequestHandler of all Backend modules
detects where a route parameter is given and
then resolves to a controller which inherits the ControllerInterface
introduced with PSR-7, and checks for a valid token.

See http://wiki.typo3.org/Blueprints/BackendRouting
for implementation details.

Resolves: #65493
Releases: master
Change-Id: I39257df45b177793c5e8f57970b4088183b78c73
Reviewed-on: http://review.typo3.org/38095
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Susanne Moog <typo3@susannemoog.de>
Tested-by: Susanne Moog <typo3@susannemoog.de>
Reviewed-by: Daniel Maier <dani-maier@gmx.de>
Tested-by: Daniel Maier <dani-maier@gmx.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>