[FEATURE] Add TCA type image_manipulation
TCA type image_manipulation brings a image manipulation wizard
to the core.

This first version brings image cropping with the possibility to
set a certain aspect ratio for the cropped area. The
sys_file_reference.crop property is extended and can now also hold
a json string to describe the image manipulation.

The LocalCropScaleMaskHelper that is used by the core
to create adjusted images is also adjusted to handle the new format.

Overriding TCA by TSConfig will be done in a followup.

Resolves: #65585
Releases: master
Change-Id: I58dee33e0f884ba2907259e8b03254c43f4c9186
Reviewed-on: http://review.typo3.org/37622
Reviewed-by: Benjamin Kott <info@bk2k.info>
Tested-by: Benjamin Kott <info@bk2k.info>
Reviewed-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Tested-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Reviewed-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Tested-by: Mathias Schreiber <mathias.schreiber@wmdb.de>
Reviewed-by: Frans Saris <franssaris@gmail.com>
Tested-by: Frans Saris <franssaris@gmail.com>