[BUGFIX] Show correct timezone when displaying FormEngine time field
"time" and "timesec" fields currently do not store time as UTC in database,
however, in the FormEngine modifications time was also handled
as being stored in UTC which results in time fields being stored and
then being modified with +/- x hours, depending on server time.
This patch reverts the behaviour back to store times as server-time.

However, in the future it would be good to store those values as UTC too,
but that will require more changes and would also be a breaking change.

Change-Id: I6c299bfa3d0387081f00c6cba7723e10236fabf0
Resolves: #65976
Releases: master
Reviewed-on: http://review.typo3.org/38112
Reviewed-by: Markus Klein <klein.t3@reelworx.at>
Reviewed-by: Jan Helke <typo3@helke.de>
Reviewed-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Tested-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>