[BUGFIX] BackendUtility: Skip rendering thumbnails for anything but files
If the stored value in a group field with type file/file_reference is a
folder the thumbCode() method skips rendering a thumbnail for that
resource. This avoids fatal error as the interface of e.g. a Folder is
different to that of a File, missing the ::isMissing() method.

Change-Id: Iedf4b5e5b17d0310811bb862c131cd70e988e72f
Resolves: #66045
Releases: master, 6.2
Reviewed-on: https://review.typo3.org/38240
Reviewed-by: Frank Nägler <frank.naegler@typo3.org>
Reviewed-by: Daniel Goerz <ervaude@gmail.com>
Tested-by: Daniel Goerz <ervaude@gmail.com>
Reviewed-by: Morton Jonuschat <m.jonuschat@mojocode.de>
Tested-by: Morton Jonuschat <m.jonuschat@mojocode.de>