[TASK] Bower component & javascript inclusion cleanup
The bower components needed for building the development
setup are now downloaded under Build/bower_components,
so no adaption is needed in the bower configuration anymore.

This makes the directory "typo3/" to only contain the necessary
files to deploy the sources at a development machine, as the required
files are copied to typo3/sysext/core/Resources/Public/JavaScript/Contrib
and commited.

Additionally, a README.txt is put in the Contrib/ directory to notify
developers that the files should not be referenced manually.

The prototype.js and bootstrap.js libraries are moved to the Contrib/
directory and their references are changed accordingly.

Resolves: #66108
Releases: master
Change-Id: Iaa0554779a317a4dcd4eb853e86a8e6459dffe58
Reviewed-on: http://review.typo3.org/38319
Reviewed-by: Daniel Sattler <sattler@b13.de>
Tested-by: Daniel Sattler <sattler@b13.de>
Reviewed-by: Andreas Fernandez <andreas.fernandez@aspedia.de>
Tested-by: Andreas Fernandez <andreas.fernandez@aspedia.de>