[BUGFIX] Prevent root folder listing for users
If a user hasn't any file mount defined or the defined file mounts
don't exist the root folder of the storage is shown. To prevent
disallowed listing of folders only admin users are allowed see and
browse root folder and editors get an information that no file mounts
are configured.

Releases: master, 6.2
Resolves: #66687
Change-Id: I301d05773f10885351034dae2b0bbd16ac20ac55
Reviewed-on: http://review.typo3.org/39089
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>