[!!!][FEATURE] BE Login form API
With the new login screen (introduced in 7.2) the possibility to
override or extend the login template was removed.
The old globals ``$GLOBALS['TBE_STYLES']['htmlTemplates']`` was
removed without replacement.

With this patch we introduce a new way to extend the login form and
add a BE login form API.

Resolves: #66669
Releases: master
Change-Id: I36ab4cdabbab55f370d1fd19b7212cf2f858db57
Reviewed-on: http://review.typo3.org/39234
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Frank Nägler <typo3@naegler.net>
Tested-by: Frank Nägler <typo3@naegler.net>
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>