[FEATURE] Allow module menu items to use sprite icons
In the course of the 7.x development, the main modules (web / file etc.)
have received a hard-coded list of fontawesome sprite icons. This way
it is not (easily) possible to override the icons and to give a custom
main module its own custom sprite icon.

A new property in the module configuration "icon" allows now to
hand in a sprite icon instead of a path to png/gif file.

At the same time the property handling for the "icon" property
inside the module menu is now simplified as a string instead
of an array with all additional information as everything is already
compiled together in the BackendModuleRepository.

All main modules shipped with the core now have proper icons
mapped inside ext_tables.php in the core.

Resolves: #66822
Releases: master
Change-Id: If1a62528d9de56034a0330c0051dfcb66e59469c
Reviewed-on: http://review.typo3.org/39316
Reviewed-by: Andreas Fernandez <typo3@scripting-base.de>
Tested-by: Andreas Fernandez <typo3@scripting-base.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>