[TASK] FormEngine: Simplify element name resolving
Data submitted by FormEngine to DataHandler is prefixed in four
different ways:
* data - Transferred data
* cmd - A dataHandler command to execute
* data_files - Contains uploaded files
* action[control] - Overwrite fields data for NULL handling

These four prefixes are given along from FormEngine to the
whole container and element structure and never change.

To simplify this stuff, the patch removes the prefixes and
hardcodes them, and introduces a new globalOptions
$elementBaseName that contains the stuff behind the prefixes.
Quite some methods can be simplified this way and become
more obvious. Usage of $elementBaseName will be extended
with the next patch.

Change-Id: I8103b1d2068382e0b01bbe5d9bd79aeba2d00b08
Resolves: #66894
Releases: master
Reviewed-on: http://review.typo3.org/39462
Reviewed-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Tested-by: Anja Leichsenring <aleichsenring@ab-softlab.de>
Reviewed-by: Wouter Wolters <typo3@wouterwolters.nl>
Tested-by: Wouter Wolters <typo3@wouterwolters.nl>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>